@@ -48,16 +48,16 @@ Coroutine\run(function () use ($config) {
     $poolSize = $GLOBALS['argv'][1] ?? 20;
     //第一个参数是并发数
     $msgNum = $GLOBALS['argv'][2] ?? 1000;
-
+    
     $GLOBALS['totalSubNum'] = $poolSize * $msgNum;
-
+    
     //第二个参数是单个发送条数
     $GLOBALS['totalNum'] = 2 * $poolSize * $msgNum;
-
+    
     $GLOBALS['startTime'] = time();
-
+    
     $pool = [];
-
+    
     for ($i = 0; $i < $poolSize; $i++) {
         $smpp = new SMPP3Client(
             [
@@ -71,40 +71,41 @@ Coroutine\run(function () use ($config) {
                 //例如每秒100，会分成10分，100ms最多发10条，如果前10ms就发送完了10条，submit的时候会自动Co sleep 90ms。
                 'fee_type'             => '01',        //资费类别
                 'client_type'          => 1,
+                'system_type'          => 'logic',
+                'mode'                 => 'transceiver', //transceiver，transmitter:receiver, transmitter, receiver，四种模式
             ]
         );
-
+        
         $arr = $smpp->login($config['server_addr'], $config['server_port'], $config['account'], $config['password'], 10); //10s登录超时
-
+        
         if ($arr !== false && $arr['command_status'] == 0) {
             var_dump('客户端' . $i . '：登陆成功，' . time());
-
+            
             $pool[] = $smpp;
-
             Swoole\Coroutine::create(function () use ($smpp, $i) {
                 while (true) {
                     //默认-1永不超时 直到有数据返回；
                     //只会收到submit回执包 或 delivery的请求包
-                    $pack = $smpp->recv(-1);
+                    $pack = $smpp->clientTransceiverListen(-1);
                     echo "Client Receive Pack: " . json_encode($pack, JSON_PRETTY_PRINT | JSON_UNESCAPED_UNICODE) . ".\n";
                     if ($pack === false) {
-                        if ($smpp->errCode === 8) {
+                        if ($smpp->getTransceiver()->getErrCode() === 8) {
                             var_dump('客户端' . $i . '：连接断开');
                             break;
                         } else {
                             continue;
                         }
                     }
-
+                    
                     if ($pack['command_id'] == SMPP3Protocol::SUBMIT_SM_RESP) {
                         @$GLOBALS['totalSubRespCnt']++;
                         if ($GLOBALS['totalSubRespCnt'] == $GLOBALS['totalSubNum']) {
                             $diff = time() - $GLOBALS['startTime'];
                             var_dump("发送结束 sub_resp 耗时 $diff");
                         }
-
+                        
                         $GLOBALS['submitTime'][time()] = ($GLOBALS['submitTime'][time()] ?? 0) + 1;
-
+                        
                         if ($pack['command_status'] === 0) {
                             $GLOBALS['totalNum']--;
                         } else {
@@ -113,20 +114,20 @@ Coroutine\run(function () use ($config) {
                         }
                     } elseif ($pack['registered_delivery'] === 1) {
                         @$GLOBALS['simuCnt']++;
-
+                        
                         @$GLOBALS['totalRepCnt']++;
-
+                        
                         if ($GLOBALS['totalRepCnt'] == $GLOBALS['totalSubNum']) {
                             $diff = time() - $GLOBALS['startTime'];
                             var_dump("发送结束 report 耗时 $diff");
                         }
-
-
+                        
+                        
                         $GLOBALS['reportTime'][time()] = ($GLOBALS['reportTime'][time()] ?? 0) + 1;
                         $GLOBALS['totalNum']--;
                         $GLOBALS['report_err'][$pack['short_message']['stat']] = 1;
                     }
-
+                    
                     if ($GLOBALS['totalNum'] <= 0) {
                         $GLOBALS['mark'] = true;
                     }
@@ -135,62 +136,62 @@ Coroutine\run(function () use ($config) {
         } else {
             var_dump('客户端' . $i . '：登陆失败');
             echo "Receive Data: " . json_encode($arr, JSON_PRETTY_PRINT | JSON_UNESCAPED_UNICODE) . ".\n";
-
+            
             foreach ($pool as $cmpp) {
                 $cmpp->logout();
             }
         }
     }
-
-
+    
+    
     Coroutine::create(function () {
         while (true) {
             var_dump('simulatorCnt:' . @$GLOBALS['simuCnt']);
             Coroutine::sleep(1);
         }
     });
-
+    
     Coroutine::create(function () use ($pool) {
         while (true) {
             if ($GLOBALS['mark']) {
                 foreach ($pool as $cmpp) {
                     $cmpp->logout();
                 }
-
+                
                 var_dump('submit response 返回分布时间分布：');
-
+                
                 foreach ($GLOBALS['submitTime'] as $time => $num) {
                     var_dump($time . ':' . $num);
                 }
-
+                
                 var_dump('report 返回分布时间分布：');
-
+                
                 foreach ($GLOBALS['reportTime'] as $time => $num) {
                     var_dump($time . ':' . $num);
                 }
-
+                
                 if (!empty($GLOBALS['resp_err'])) {
                     var_dump('response错误：');
-
+                    
                     var_dump(array_keys($GLOBALS['resp_err']));
                 }
-
+                
                 if (!empty($GLOBALS['report_err'])) {
                     var_dump('report错误：');
                     echo "Receive Data: " . json_encode($GLOBALS['report_err'], JSON_PRETTY_PRINT | JSON_UNESCAPED_UNICODE) . ".\n";
                     // var_dump(array_keys($GLOBALS['report_err']));
                 }
-
+                
                 break;
             }
-
+            
             Coroutine::sleep(1);
         }
     });
-
-    $mobile = $config['mobile'];
+    
+    $mobile     = $config['mobile'];
     $msgContent = $config['msg'];
-
+    
     foreach ($pool as $smpp) {
         Coroutine::create(function () use ($smpp, $msgNum, &$mobile, $msgContent) {
             $s = 0;
@@ -198,7 +199,7 @@ Coroutine\run(function () use ($config) {
                 $smpp->submit((string)$mobile, $msgContent, '0615', -1, ($s++) % 255); //默认-1 永不超时
                 $mobile++;
             }
-
+            
             var_dump('发送完毕，' . time());
         });
     }
